refactor(core): better layers separation
