--- ktor-service/src/main/kotlin/Frameworks.kt
@@ -1,19 +1,0 @@
-package com.pavelryzh
-
-import io.ktor.server.application.*
-import org.koin.dsl.module
-import org.koin.ktor.plugin.Koin
-import org.koin.logger.slf4jLogger
-
-fun Application.configureFrameworks() {
-    install(Koin) {
-        slf4jLogger()
-        modules(module {
-            single<HelloService> {
-                HelloService {
-                    println(environment.log.info("Hello, World!"))
-                }
-            }
-        })
-    }
-}
