--- ktor-service/src/main/kotlin/HelloService.kt
@@ -1,5 +1,0 @@
-package com.pavelryzh
-
-fun interface HelloService {
-    fun sayHello()
-}
