--- ktor-service/src/main/kotlin/Routing.kt
@@ -1,40 +1,0 @@
-package com.pavelryzh
-
-import com.auth0.jwt.JWT
-import com.auth0.jwt.algorithms.Algorithm
-import io.github.flaxoos.ktor.server.plugins.kafka.Kafka
-import io.github.flaxoos.ktor.server.plugins.kafka.MessageTimestampType
-import io.github.flaxoos.ktor.server.plugins.kafka.TopicName
-import io.github.flaxoos.ktor.server.plugins.kafka.admin
-import io.github.flaxoos.ktor.server.plugins.kafka.common
-import io.github.flaxoos.ktor.server.plugins.kafka.consumer
-import io.github.flaxoos.ktor.server.plugins.kafka.consumerConfig
-import io.github.flaxoos.ktor.server.plugins.kafka.consumerRecordHandler
-import io.github.flaxoos.ktor.server.plugins.kafka.producer
-import io.github.flaxoos.ktor.server.plugins.kafka.registerSchemas
-import io.github.flaxoos.ktor.server.plugins.kafka.topic
-import io.ktor.client.HttpClient
-import io.ktor.http.*
-import io.ktor.openapi.*
-import io.ktor.serialization.kotlinx.json.*
-import io.ktor.server.application.*
-import io.ktor.server.auth.*
-import io.ktor.server.auth.jwt.*
-import io.ktor.server.plugins.contentnegotiation.*
-import io.ktor.server.plugins.cors.routing.*
-import io.ktor.server.plugins.forwardedheaders.*
-import io.ktor.server.plugins.openapi.*
-import io.ktor.server.plugins.swagger.*
-import io.ktor.server.response.*
-import io.ktor.server.routing.*
-import org.koin.dsl.module
-import org.koin.ktor.plugin.Koin
-import org.koin.logger.slf4jLogger
-
-fun Application.configureRouting() {
-    routing {
-        get("/") {
-            call.respondText("Hello World!")
-        }
-    }
-}
