+++ ktor-service/src/main/kotlin/plugins/Frameworks.kt
@@ -1,0 +1,15 @@
+package plugins
+
+import com.pavelryzh.service.TrafficService
+import io.ktor.server.application.*
+import org.koin.dsl.module
+import org.koin.ktor.plugin.Koin
+import org.koin.logger.slf4jLogger
+
+//fun Application.configureFrameworks() {
+//    install(Koin) {
+//        slf4jLogger()
+//        modules(module {
+//        })
+//    }
+//}
