+++ ktor-service/src/main/kotlin/plugins/Security.kt
@@ -1,4 +1,4 @@
-package com.pavelryzh
+package com.pavelryzh.plugins
 
 import com.auth0.jwt.JWT
 import com.auth0.jwt.algorithms.Algorithm
